Move abstract KeyboardEvent to internal package
