+++ flow-server/src/main/java/com/vaadin/flow/component/KeyEventListener.java
@@ -18,6 +18,8 @@ package com.vaadin.flow.component;
 import java.util.Arrays;
 import java.util.EnumSet;
 
+import com.vaadin.flow.component.internal.KeyboardEvent;
+
 /**
  * A conditional event listener for {@link KeyboardEvent}s.
  *
