+++ flow-server/src/main/java/com/vaadin/flow/component/KeyUpEvent.java
@@ -15,6 +15,8 @@
  */
 package com.vaadin.flow.component;
 
+import com.vaadin.flow.component.internal.KeyboardEvent;
+
 /**
  * The event when a key is released.
  *
